3차 병합

+++ src/main/resources/static/js/calendar/calendar.js
@@ -10,6 +10,8 @@ let local_date_end;
 let clickedDate;				// 클릭한 날짜의 값을 저장해 두는것임.
 let id;									// form 의 번호임.
 
+const team_id = $("meta[name='team_id']").attr('content');
+const tm_grade = $("meta[name='tm_grade']").attr('content');
 const header = $("meta[name='_csrf_header']").attr('content');
 const token = $("meta[name='_csrf']").attr('content');
 
@@ -64,7 +66,7 @@ function createCalendar(month, year) {
 		if (i === currentDate.getDate() && month === currentDate.getMonth() + 1 && year === currentDate.getFullYear()) {
 			day.classList.add('today');
 		}
-		
+
 	}
 	$(document).off("click");
 
@@ -84,11 +86,15 @@ function createCalendar(month, year) {
 
 		local_date_start = clickDate.toISOString().slice(0, 16);
 		local_date_end = clickDateSec.toISOString().slice(0, 16);
-
+		let data = {
+			"date": clickedDate,
+			"team_id": team_id,
+			"tm_grade": tm_grade
+		};
 		$.ajax({
 			url: "schedule",
 			type: "get",
-			data: { "date": clickedDate },
+			data: data,
 			success: function(html) {
 
 				$(".schedule").empty();
@@ -152,6 +158,7 @@ window.onload = function() {
 
 // 여기는 버튼 클릭시 글작성 div 왔다 갔다 하는 스크립트
 function btnClick() {
+
 	const mydiv = document.getElementById('my-div');
 
 	if (mydiv.style.display !== 'block') {
@@ -224,6 +231,10 @@ function writeSchedule() {
 		return;
 	}
 
+	if (team_id != null) {
+		$("#form").append($('<input>', { type: 'hidden', name: 'team_id', val: team_id }));
+	}
+
 	$.ajax({
 		url: 'schedule',
 		type: 'POST',
@@ -246,16 +257,35 @@ function writeSchedule() {
 // 삭제 버튼
 function deleteScheduel(btn) {
 	const schedule_id = btn.value;
-	$.ajax({
-		url: 'schedule/del',
-		type: 'delete',
-		data: { schedule_id: schedule_id },
-		beforeSend: function(xhr) {
-			xhr.setRequestHeader(header, token);
-		},
-		success: function() {
-			$(".schedule").load(window.location.href + "");
-			scheduleAjax();
+	Swal.fire({
+		title: "삭제 하시겠습니까?",
+		icon:"question",
+		showCancelButton: true,
+		confirmButtonText: "네",
+		denyButtonText: "아니요"
+	}).then((result) => {
+		if (result.isConfirmed) {
+			Swal.fire({
+				title: "삭제 되었습니다.",
+				icon: "success"
+			});
+			$.ajax({
+				url: 'schedule/del',
+				type: 'delete',
+				data: { schedule_id: schedule_id },
+				beforeSend: function(xhr) {
+					xhr.setRequestHeader(header, token);
+				},
+				success: function() {
+					$(".schedule").load(window.location.href + "");
+					scheduleAjax();
+				}
+			});
+		} else {
+			Swal.fire({
+				title: "삭제되지 않았습니다.",
+				icon: "error"
+			});
 		}
 	});
 }
@@ -289,10 +319,15 @@ function editSchedule(btn) {
 
 // 처리 이후 페이지 새로고침 하는 ajax;
 function scheduleAjax() {
+	let data = {
+		"date": clickedDate,
+		"team_id": team_id,
+		"tm_grade": tm_grade
+	};
 	$.ajax({
 		url: "schedule",
 		type: "GET",
-		data: { "date": clickedDate },
+		data: data,
 		success: function(html) {
 			$(".schedule").empty();
 			$(".schedule").append(html);
